refactor(user-validation): enhance null checks and simplify validation logic
- Moved null checks to a dedicated `checkNulls` method for better readability.
- Added `UserFieldNullException` to handle cases with missing user fields.
- Updated validation logic to streamline field checks and improve maintainability.
